pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deeptrade"
version = "0.0.1"
description = "LLM-driven A-share (Shanghai/Shenzhen main board) stock screening CLI"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "DeepTrade" }]
license = { text = "MIT" }
keywords = ["stock", "a-share", "llm", "tushare", "deepseek", "duckdb", "cli"]

dependencies = [
    "typer>=0.12",
    "questionary>=2.0",
    "rich>=13.7",
    "duckdb>=1.0",
    "tushare>=1.4",
    "openai>=1.0",
    "pydantic>=2.7",
    "tenacity>=8.0",
    "pyyaml>=6.0",
    "keyring>=25.0",
    "pandas>=2.2",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "ruff>=0.5",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "types-PyYAML>=6.0",
]

[project.scripts]
deeptrade = "deeptrade.cli:app"

[project.urls]
Documentation = "https://github.com/example/deeptrade"
Repository   = "https://github.com/example/deeptrade"

[tool.hatch.build.targets.wheel]
packages = ["deeptrade"]

[tool.hatch.build.targets.wheel.force-include]
"deeptrade/core/migrations" = "deeptrade/core/migrations"

[tool.hatch.build.targets.sdist]
include = ["/deeptrade", "/tests", "/README.md", "/DESIGN.md", "/PLAN.md"]
